feat(renderer): add image support with per render workdirs (#103)
- added image support to renderer
- assign compile workdir based on renderId
- fixed filepaths and permissions for images
- clean up workdir after compilation
- updated README
- added log statements to api for requests that failed silently, namely
  ongoing compilations and requests that failed thrice without change.
- added check to prevent multiple images of the same document having the same user-supplied-name.

#87

+++ src/api/src/main/java/com/latech/api/api/ImageController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.nio.file.FileAlreadyExistsException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -69,10 +70,14 @@ public class ImageController {
                         file.getOriginalFilename(),
                         file));
             }
-        } catch (IOException e) {
+        } catch ( FileAlreadyExistsException e){
+            log.error("Exception while renaming - filename already exists. ", e);
+            return ResponseEntity.status( 409 ).body("Error while uploading: Filename already exists");
+        } catch ( IOException e) {
             log.error("Exception while uploading file", e);
             return ResponseEntity.internalServerError().body("Error while uploading file");
         }
+
         List<DocumentImageDto> uploadedFiles = entities.stream()
                 .map(entity -> DocumentImageDto.builder()
                         .id(entity.getImageId())
@@ -95,10 +100,10 @@ public class ImageController {
             return ResponseEntity.badRequest().body(null);
         }
 
-        Optional<DocumentImage> picture = documentImageService.getPictureFromDocumentAndImageId(
+        Optional<DocumentImage> documentImage = documentImageService.getPictureFromDocumentAndImageId(
                 docId, imageId);
 
-        if (picture.isEmpty()) {
+        if (documentImage.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
 
@@ -108,7 +113,7 @@ public class ImageController {
             return ResponseEntity.notFound().build();
         }
 
-        MediaType mediaType = MediaType.parseMediaType(picture.get().getMimeType());
+        MediaType mediaType = MediaType.parseMediaType(documentImage.get().getMimeType());
 
         return ResponseEntity.ok()
                 .contentType(mediaType)
@@ -128,6 +133,10 @@ public class ImageController {
 
         if (ObjectUtils.isEmpty(imageId)) {
             return ResponseEntity.badRequest().body(null);
+        }
+
+        if (this.documentImageService.pictureExistsWithDocumentIdAndImageName( docId, file.getName() )){
+            return ResponseEntity.status(409).body( "Failed to rename image, a file with that name already exists." );
         }
 
         DocumentImage updatedImage = this.documentImageService.updatePicture(docId, imageId, file.getName());
@@ -175,7 +184,10 @@ public class ImageController {
             return ResponseEntity.badRequest().body("Image ID is empty");
         }
 
-        Boolean deleted = this.imageService.deleteImage(docId, imageId);
+        boolean deleted = false;
+        if (this.documentImageService.pictureExistsWithDocumentIdAndImageId( docId, imageId )) {
+            deleted = this.imageService.deleteImage( docId, imageId );
+        }
 
         if (!deleted) {
             return ResponseEntity.status(500).body("Failed to delete image");
